added roommdatabase

+++ app/src/main/java/com/taru/data/remote/ip/dto/IpDto.kt
@@ -6,6 +6,6 @@ import com.squareup.moshi.JsonClass
  * Created by Niraj on 16-01-2023.
  */
 @JsonClass(generateAdapter = true)
-data class IpDto(var query: String, var lat: Double, var lon: Double){
+data class IpDto(var query: String, var lat: Float, var lon: Float){
 
 }
