+++ app/src/main/java/com/taru/data/remote/ip/mapper/DtoMapper.kt
@@ -5,6 +5,6 @@ import com.taru.data.remote.ip.dto.IpDto
 /**
  * Created by Niraj on 16-01-2023.
  */
-fun IpDto.toLatLng(): Array<Double> {
+fun IpDto.toLatLng(): Array<Float> {
     return arrayOf(lat, lon)
 }
